+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/FastInitBoxTestGenerated.java
@@ -280,6 +280,12 @@ public class FastInitBoxTestGenerated extends AbstractFastInitBoxTest {
       }
 
       @Test
+      @TestMetadata("explicitMultibindingType.kt")
+      public void testExplicitMultibindingType() {
+        runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/explicitMultibindingType.kt");
+      }
+
+      @Test
       @TestMetadata("exposeImplBinding.kt")
       public void testExposeImplBinding() {
         runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/exposeImplBinding.kt");